Netflix_case_study
We are given the dataset of one of the leading streaming platforms which has over 10000 moves and TV shows available on their platform as of mid-2021 with over 22M subscribers around the globe. This tabular dataset consists of listings of all the movies and tv shows available on Netflix, along with details such as - cast, directors, ratings, release year, duration, etc.
Analysis of this data is done and valuable insights and recommendations are generated which will help in deciding which type of shows/movies to produce and how they can grow the business in different countries.